Recreated the bug we multiple parameters colliding and overriding the request body example provided by SwaggerRequestExample.  In this case because we had a header that was of type string and one of the examples in the example repository was of type string and ServiceProviderExamplesOperationFilter is setup to run after ExamplesOperationFilter in SwaggerGenOptionsExtensions.ExampleFilters; the last parameter in the parameterDescription list with a matching IExamplesProvider clobbers any existing example that was set by ExamplesOperationFilter even if it is not of type BindingSource.Body.
Suggested fix is to check the Source parameter on the parameterDescription and only Set the RequestBody example if it matches.  It may also be prudent to make ExamplesOperationFilter run after ServiceProviderExamplesOperationFilter, that way the last example value that gets set is the one manually specified via the attributes, but I'm not sure if you specifically intended for ServiceProviderExamplesOperationFilter to run after ExamplesOperationFilter or not.

There may be a related bug in the SetResponseExamples function, but I haven't explored that.